bd18398: Open only few of the most helpful tabs.
The IDE remembers open tabs. Try guessing which tabs are the most
helpful when trying to understand the EVK control flow and which
are the most important in order to understand the LED control.

I guess this depends on the person reading code - we all probably
have our own style of diving into the code. But let's try.

Open few most crucial code-files in IDE tabs when project is opened.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>