bd18398: Clear protection LED when error condition is cleared
The detection of SHORT error toggles protection LED 'constantly on'
as was designed. When SHORT error is cleared at IC - the LED status
is not changed.

Cache LED error statuses and check if the errors were cleared. Free
the protection LED to indicate other errors (or no errors) if SHORT
is cleared - or if SHORT is not active and OPEN was cleared.

Fixes: a2621d6898e7 ("bd18398-demo: Fix board LED usage")
Reported-by: Thady Bruton <thady.bruton@de.rohmeurope.com>
Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>