bd18398-demo: Make DEMO ALIVE prints less frequent
Demo may be spilling out the "DEMO ALIVE" prints quite
frequently. Increase the amount of loops when we do not print.
Especially the LIMP_HOME mode when we omit the SPI traffic makes
the "DEMO ALIVE" noisy. It would be better to use some more
reliably timing for the print here - but let's just make it less
frequent now.

Also, change the loop index variable to unsigned for better defined
overflow.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>